+++ activerecord/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Fix context of *_through association scope from through scope to target scope (as in other association)
+
+    *Ivan Antropov*
+
 *   Bumped the minimum supported version of PostgreSQL to >= 9.1.
     Both PG 9.0 and 8.4 are past their end of life date:
     http://www.postgresql.org/support/versioning/